backend/app/config/security.py
from passlib.context import CryptContext
from jose import JWTError, jwt

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

def encrypt_data(data: str, key: str) -> str:
    return Fernet(key).encrypt(data.encode()).decode()

def verify_password(plain_password: str, hashed_password: str):
    return pwd_context.verify(plain_password, hashed_password)