backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.3
python-binance==1.0.20
motor==3.3.3
pydantic==2.4.2
python-jose[cryptography]==3.3.1
passlib[bcrypt]==1.7.5
cryptography==41.0.0
python-dotenv==1.0.1
pytest==7.4.1